Move Monitored out of wrappers __init__
